Add per column selection handling logic
For #7

- added columns_checked signal to the view.checkableheaderview which
emits its id and a list of checked columns

- added controller.update_checked_columns callback function (passed from
controller into view._create_ui_table where it is connected as a slot to
the view.checkableheader.columns_checked signal) which updates the model
with the checked columns

- added a ProcessedTableManager class in the model to handle
ProcessedTable objects, which are used to store the checked columns (and
the overall checked status) for each processed table. This is kept
separate from the (now renamed) ProcessedTableDBEntry (which was
previously mislabelled ProcessedTableWrapper) that stores the actual
data in a db. So we keep the light metadata in an in-memory object and
the heavy data in the db

- added a num_columns parameter to the ProcessedTable class, which is
used at instantiation to populate the checked columns list with all the
columns checked by default (we use
`checked_columns = list(range(num_columns)`)

- updated the model.update_processed_tables() method to compute the
number of columns in the processed table and pass it to the
ProcessedTable constructor

- modified view._create_ui_table() to set the table id in the
checkableheaderview object